added forever example
using forever is simple, just point it to a file, and start it. It
monitors the process and if it exits it will spawn a new process.
Node.JS Cluster API works similar but you can use it to spawn new
express applications for a more graceful solution. The cluster api can
shutdown workers, and there is zero-downtime because workers are
resuscitated and ready workers take their places while the others come
back to life.